+++ src/XlsxWorkSheet.cpp
@@ -3,13 +3,13 @@
 using namespace Rcpp;
 
 // [[Rcpp::export]]
-void xlsx_cells(std::string path, int sheet = 0) {
-  return XlsxWorkSheet(path, sheet).printCells();
+void xlsx_cells(std::string path, int sheet = 0, int nskip = 0) {
+  return XlsxWorkSheet(path, sheet, nskip).printCells();
 }
 
 // [[Rcpp::export]]
-IntegerVector xlsx_dim(std::string path, int sheet = 0) {
-  XlsxWorkSheet ws(path, sheet);
+IntegerVector xlsx_dim(std::string path, int sheet = 0, int nskip = 0) {
+  XlsxWorkSheet ws(path, sheet, nskip);
   return IntegerVector::create(ws.nrow(), ws.ncol());
 }
 
@@ -25,7 +25,7 @@ CharacterVector xlsx_col_types(std::string path, int sheet = 0,
                                CharacterVector na = CharacterVector(), int nskip = 0,
                                int n = 100) {
 
-  XlsxWorkSheet ws(path, sheet);
+  XlsxWorkSheet ws(path, sheet, nskip);
   std::vector<CellType> types = ws.colTypes(na, nskip, n);
 
   CharacterVector out(types.size());
@@ -38,14 +38,14 @@ CharacterVector xlsx_col_types(std::string path, int sheet = 0,
 
 // [[Rcpp::export]]
 CharacterVector xlsx_col_names(std::string path, int sheet = 0, int nskip = 0) {
-  return XlsxWorkSheet(path, sheet).colNames(nskip);
+  return XlsxWorkSheet(path, sheet, nskip).colNames();
 }
 
 // [[Rcpp::export]]
 List read_xlsx_(std::string path, int sheet, RObject col_names,
                 RObject col_types, std::vector<std::string> na, int nskip = 0) {
 
-  XlsxWorkSheet ws(path, sheet);
+  XlsxWorkSheet ws(path, sheet, nskip);
 
   if (ws.nrow() == 0 && ws.ncol() == 0) {
     return Rcpp::List(0);
@@ -62,7 +62,7 @@ List read_xlsx_(std::string path, int sheet, RObject col_names,
   {
     sheetHasColumnNames = as<bool>(col_names);
     if (sheetHasColumnNames) {
-      colNames = ws.colNames(std::max(ws.min_row(), nskip));
+      colNames = ws.colNames();
     } else {
       int p = ws.ncol();
       colNames = CharacterVector(p);
@@ -76,16 +76,11 @@ List read_xlsx_(std::string path, int sheet, RObject col_names,
     Rcpp::stop("`col_names` must be a logical or character vector");
   }
 
-  // if we know there's no data reading to be done, we should return a
-  // 0 row tibble HERE with the correct colNames
-  // TO THINK: case where colTypes were explicitly specified but no data to read
-
   // Standardise column types --------------------------------------------------
   std::vector<CellType> colTypes;
   switch(TYPEOF(col_types)) {
   case NILSXP:
-    colTypes = ws.colTypes(na, std::max(ws.min_row(), nskip),
-                           100, sheetHasColumnNames);
+    colTypes = ws.colTypes(na, 100, sheetHasColumnNames);
     break;
   case STRSXP:
     colTypes = cellTypes(as<CharacterVector>(col_types));
@@ -94,6 +89,5 @@ List read_xlsx_(std::string path, int sheet, RObject col_names,
     Rcpp::stop("`col_types` must be a character vector or NULL");
   }
 
-  return ws.readCols(colNames, colTypes, na,
-                     std::max(ws.min_row(), nskip), sheetHasColumnNames);
+  return ws.readCols(colNames, colTypes, na, sheetHasColumnNames);
 }